Arrange project for multiple runtime versions
