+++ DxLibSpineViewer.sln
@@ -1,0 +1,81 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.11.35219.272
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "DxLibSpineViewerC-3.6", "projects\DxLibSpineViewerC-3.6\DxLibSpineViewerC-3.6.vcxproj", "{081CCF8D-2041-483C-8BA0-7CBF3083F3A5}"
+EndProject
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "DxLibSpineViewer-3.8", "projects\DxLibSpineViewer-3.8\DxLibSpineViewer-3.8.vcxproj", "{8F3A1A04-940D-4D5D-97E5-82C26B822D44}"
+EndProject
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "DxLibSpineViewer-4.0", "projects\DxLibSpineViewer-4.0\DxLibSpineViewer-4.0.vcxproj", "{AC0252A3-141A-4E49-9CC3-D074E5413DB4}"
+EndProject
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "DxLibSpineViewerC-3.5", "projects\DxLibSpineViewerC-3.5\DxLibSpineViewerC-3.5.vcxproj", "{81E05375-9432-4A0B-8DF1-9E1C01726F09}"
+EndProject
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "DxLibSpineViewerC-3.7", "projects\DxLibSpineViewerC-3.7\DxLibSpineViewerC-3.7.vcxproj", "{6E1168E1-446A-4267-ADDC-3A831BD8B080}"
+EndProject
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "DxLibSpineViewerC-2.1", "projects\DxLibSpineViewerC-2.1\DxLibSpineViewerC-2.1.vcxproj", "{7AB026B4-06D1-4F34-AB67-10DC5B5B4E69}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{081CCF8D-2041-483C-8BA0-7CBF3083F3A5}.Debug|x64.ActiveCfg = Debug|x64
+		{081CCF8D-2041-483C-8BA0-7CBF3083F3A5}.Debug|x64.Build.0 = Debug|x64
+		{081CCF8D-2041-483C-8BA0-7CBF3083F3A5}.Debug|x86.ActiveCfg = Debug|Win32
+		{081CCF8D-2041-483C-8BA0-7CBF3083F3A5}.Debug|x86.Build.0 = Debug|Win32
+		{081CCF8D-2041-483C-8BA0-7CBF3083F3A5}.Release|x64.ActiveCfg = Release|x64
+		{081CCF8D-2041-483C-8BA0-7CBF3083F3A5}.Release|x64.Build.0 = Release|x64
+		{081CCF8D-2041-483C-8BA0-7CBF3083F3A5}.Release|x86.ActiveCfg = Release|Win32
+		{081CCF8D-2041-483C-8BA0-7CBF3083F3A5}.Release|x86.Build.0 = Release|Win32
+		{8F3A1A04-940D-4D5D-97E5-82C26B822D44}.Debug|x64.ActiveCfg = Debug|x64
+		{8F3A1A04-940D-4D5D-97E5-82C26B822D44}.Debug|x64.Build.0 = Debug|x64
+		{8F3A1A04-940D-4D5D-97E5-82C26B822D44}.Debug|x86.ActiveCfg = Debug|Win32
+		{8F3A1A04-940D-4D5D-97E5-82C26B822D44}.Debug|x86.Build.0 = Debug|Win32
+		{8F3A1A04-940D-4D5D-97E5-82C26B822D44}.Release|x64.ActiveCfg = Release|x64
+		{8F3A1A04-940D-4D5D-97E5-82C26B822D44}.Release|x64.Build.0 = Release|x64
+		{8F3A1A04-940D-4D5D-97E5-82C26B822D44}.Release|x86.ActiveCfg = Release|Win32
+		{8F3A1A04-940D-4D5D-97E5-82C26B822D44}.Release|x86.Build.0 = Release|Win32
+		{AC0252A3-141A-4E49-9CC3-D074E5413DB4}.Debug|x64.ActiveCfg = Debug|x64
+		{AC0252A3-141A-4E49-9CC3-D074E5413DB4}.Debug|x64.Build.0 = Debug|x64
+		{AC0252A3-141A-4E49-9CC3-D074E5413DB4}.Debug|x86.ActiveCfg = Debug|Win32
+		{AC0252A3-141A-4E49-9CC3-D074E5413DB4}.Debug|x86.Build.0 = Debug|Win32
+		{AC0252A3-141A-4E49-9CC3-D074E5413DB4}.Release|x64.ActiveCfg = Release|x64
+		{AC0252A3-141A-4E49-9CC3-D074E5413DB4}.Release|x64.Build.0 = Release|x64
+		{AC0252A3-141A-4E49-9CC3-D074E5413DB4}.Release|x86.ActiveCfg = Release|Win32
+		{AC0252A3-141A-4E49-9CC3-D074E5413DB4}.Release|x86.Build.0 = Release|Win32
+		{81E05375-9432-4A0B-8DF1-9E1C01726F09}.Debug|x64.ActiveCfg = Debug|x64
+		{81E05375-9432-4A0B-8DF1-9E1C01726F09}.Debug|x64.Build.0 = Debug|x64
+		{81E05375-9432-4A0B-8DF1-9E1C01726F09}.Debug|x86.ActiveCfg = Debug|Win32
+		{81E05375-9432-4A0B-8DF1-9E1C01726F09}.Debug|x86.Build.0 = Debug|Win32
+		{81E05375-9432-4A0B-8DF1-9E1C01726F09}.Release|x64.ActiveCfg = Release|x64
+		{81E05375-9432-4A0B-8DF1-9E1C01726F09}.Release|x64.Build.0 = Release|x64
+		{81E05375-9432-4A0B-8DF1-9E1C01726F09}.Release|x86.ActiveCfg = Release|Win32
+		{81E05375-9432-4A0B-8DF1-9E1C01726F09}.Release|x86.Build.0 = Release|Win32
+		{6E1168E1-446A-4267-ADDC-3A831BD8B080}.Debug|x64.ActiveCfg = Debug|x64
+		{6E1168E1-446A-4267-ADDC-3A831BD8B080}.Debug|x64.Build.0 = Debug|x64
+		{6E1168E1-446A-4267-ADDC-3A831BD8B080}.Debug|x86.ActiveCfg = Debug|Win32
+		{6E1168E1-446A-4267-ADDC-3A831BD8B080}.Debug|x86.Build.0 = Debug|Win32
+		{6E1168E1-446A-4267-ADDC-3A831BD8B080}.Release|x64.ActiveCfg = Release|x64
+		{6E1168E1-446A-4267-ADDC-3A831BD8B080}.Release|x64.Build.0 = Release|x64
+		{6E1168E1-446A-4267-ADDC-3A831BD8B080}.Release|x86.ActiveCfg = Release|Win32
+		{6E1168E1-446A-4267-ADDC-3A831BD8B080}.Release|x86.Build.0 = Release|Win32
+		{7AB026B4-06D1-4F34-AB67-10DC5B5B4E69}.Debug|x64.ActiveCfg = Debug|x64
+		{7AB026B4-06D1-4F34-AB67-10DC5B5B4E69}.Debug|x64.Build.0 = Debug|x64
+		{7AB026B4-06D1-4F34-AB67-10DC5B5B4E69}.Debug|x86.ActiveCfg = Debug|Win32
+		{7AB026B4-06D1-4F34-AB67-10DC5B5B4E69}.Debug|x86.Build.0 = Debug|Win32
+		{7AB026B4-06D1-4F34-AB67-10DC5B5B4E69}.Release|x64.ActiveCfg = Release|x64
+		{7AB026B4-06D1-4F34-AB67-10DC5B5B4E69}.Release|x64.Build.0 = Release|x64
+		{7AB026B4-06D1-4F34-AB67-10DC5B5B4E69}.Release|x86.ActiveCfg = Release|Win32
+		{7AB026B4-06D1-4F34-AB67-10DC5B5B4E69}.Release|x86.Build.0 = Release|Win32
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {F399BC77-F112-4446-8A2E-81158B35D206}
+	EndGlobalSection
+EndGlobal
