--- DxlibSpineTest.sln
@@ -1,41 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.9.34714.143
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "DxlibSpineTest", "DxlibSpineTest\DxlibSpineTest.vcxproj", "{B7A2665A-71C5-4F5A-AE16-337712FE278F}"
-EndProject
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "DxlibSpineTestC", "DxlibSpineTestC\DxlibSpineTestC.vcxproj", "{2C51758B-6C25-482D-A555-DA030AE0C3B0}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{B7A2665A-71C5-4F5A-AE16-337712FE278F}.Debug|x64.ActiveCfg = Debug|x64
-		{B7A2665A-71C5-4F5A-AE16-337712FE278F}.Debug|x64.Build.0 = Debug|x64
-		{B7A2665A-71C5-4F5A-AE16-337712FE278F}.Debug|x86.ActiveCfg = Debug|Win32
-		{B7A2665A-71C5-4F5A-AE16-337712FE278F}.Debug|x86.Build.0 = Debug|Win32
-		{B7A2665A-71C5-4F5A-AE16-337712FE278F}.Release|x64.ActiveCfg = Release|x64
-		{B7A2665A-71C5-4F5A-AE16-337712FE278F}.Release|x64.Build.0 = Release|x64
-		{B7A2665A-71C5-4F5A-AE16-337712FE278F}.Release|x86.ActiveCfg = Release|Win32
-		{B7A2665A-71C5-4F5A-AE16-337712FE278F}.Release|x86.Build.0 = Release|Win32
-		{2C51758B-6C25-482D-A555-DA030AE0C3B0}.Debug|x64.ActiveCfg = Debug|x64
-		{2C51758B-6C25-482D-A555-DA030AE0C3B0}.Debug|x64.Build.0 = Debug|x64
-		{2C51758B-6C25-482D-A555-DA030AE0C3B0}.Debug|x86.ActiveCfg = Debug|Win32
-		{2C51758B-6C25-482D-A555-DA030AE0C3B0}.Debug|x86.Build.0 = Debug|Win32
-		{2C51758B-6C25-482D-A555-DA030AE0C3B0}.Release|x64.ActiveCfg = Release|x64
-		{2C51758B-6C25-482D-A555-DA030AE0C3B0}.Release|x64.Build.0 = Release|x64
-		{2C51758B-6C25-482D-A555-DA030AE0C3B0}.Release|x86.ActiveCfg = Release|Win32
-		{2C51758B-6C25-482D-A555-DA030AE0C3B0}.Release|x86.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {4812C409-CFD8-4A1A-B7D3-1A150D6A646A}
-	EndGlobalSection
-EndGlobal
