+++ shared-src/spine_player_shared.h
@@ -2,9 +2,9 @@
 #define SPINE_PLAYER_SHARED_H_
 
 #ifdef DXLIB_SPINE_CPP
-#include "../DxlibSpineTest/dxlib_spine_player.h"
+#include "../DxLibSpineCpp/dxlib_spine_player.h"
 #elif DXLIB_SPINE_C
-#include "../DxlibSpineTestC/dxlib_spine_player_c.h"
+#include "../DxLibSpineC/dxlib_spine_player_c.h"
 using CDxLibSpinePlayer = CDxLibSpinePlayerC;
 #endif
